feat(DTFS2-7669): refactor to make middleware more DRY
