+++ libs/common/src/middleware/feature-flags/index.ts
@@ -1,1 +1,1 @@
-export * from './deal-cancellation';
+export * from './validate-feature-flag-enabled';
