--- libs/common/src/middleware/feature-flags/portal-facility-amendments.ts
@@ -1,15 +1,0 @@
-import { RequestHandler } from 'express';
-import { HttpStatusCode } from 'axios';
-import { isPortalFacilityAmendmentsFeatureFlagEnabled } from '../../helpers';
-
-/**
- * Middleware to check if the portal facility amendments feature flag is enabled
- */
-export const validatePortalFacilityAmendmentsEnabled: RequestHandler = (req, res, next) => {
-  if (!isPortalFacilityAmendmentsFeatureFlagEnabled()) {
-    console.info(`Portal facility amendments feature flag disabled, accessing ${req.originalUrl}`);
-    return res.sendStatus(HttpStatusCode.NotFound);
-  }
-
-  return next();
-};
